course teacher crud

+++ ECourseWeb/src/main/java/com/htt/pojo/Teacher.java
@@ -42,7 +42,7 @@ public class Teacher implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 100)
@@ -60,20 +60,20 @@ public class Teacher implements Serializable {
     public Teacher() {
     }
 
-    public Teacher(Integer id) {
+    public Teacher(Long id) {
         this.id = id;
     }
 
-    public Teacher(Integer id, String position) {
+    public Teacher(Long id, String position) {
         this.id = id;
         this.position = position;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
